requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.7.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.0=pypi_0
django=4.2.5=pypi_0
django-cors-headers=4.2.0=pypi_0
djangorestframework=3.14.0=pypi_0
djangorestframework-simplejwt=5.3.0=pypi_0
idna=3.4=pypi_0
libffi=3.4.4=hd77b12b_0
openssl=3.0.10=h2bbff1b_2
pip=23.2.1=py311haa95532_0
pyjwt=2.8.0=pypi_0
python=3.11.4=he1021f5_0
python-dotenv=1.0.0=pypi_0
pytz=2023.3.post1=pypi_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py311haa95532_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=pypi_0
stripe=6.7.0=pypi_0
tk=8.6.12=h2bbff1b_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
